+++ evolver/app/main.py
@@ -1,13 +1,12 @@
 import asyncio
 from contextlib import asynccontextmanager
-from typing import Any
 
-import pydantic
 from fastapi import FastAPI
 
 import evolver.util
 from evolver import __project__, __version__
-from evolver.base import BaseConfig, BaseInterface, ImportString, require_all_fields
+from evolver.app.models import SchemaResponse
+from evolver.base import ImportString, require_all_fields
 from evolver.device import Evolver
 from evolver.settings import app_settings
 
@@ -57,25 +56,6 @@ async def get_state():
 async def update_evolver(config: EvolverConfigWithoutDefaults):
     app.state.evolver = Evolver.create(config)
     app.state.evolver.config_model.save(app_settings.CONFIG_FILE)
-
-
-class SchemaResponse(pydantic.BaseModel):
-    classinfo: ImportString
-    config: dict | None = None
-    input: dict | None = None
-    output: dict | None = None
-
-    def model_post_init(self, __context: Any) -> None:
-        if issubclass(self.classinfo, BaseConfig):
-            self.config = self.classinfo.model_json_schema()
-        elif issubclass(self.classinfo, BaseInterface):
-            self.config = self.classinfo.Config.model_json_schema()
-
-            if hasattr(self.classinfo, "Input") and issubclass(self.classinfo.Input, pydantic.BaseModel):
-                self.input = self.classinfo.Input.model_json_schema()
-
-            if hasattr(self.classinfo, "Output") and issubclass(self.classinfo.Output, pydantic.BaseModel):
-                self.input = self.classinfo.Output.model_json_schema()
 
 
 @app.get("/schema/", response_model=SchemaResponse)
